% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Soma dois Elementos}
\usage{
somar(x, y)
}
\arguments{
\item{x}{primeiro elemento}

\item{y}{segundo elemento}
}
\description{
somar retorna a soma de dois valores
}
\examples{
x <- 1
y <- 2
somar(x,y)

}
